pawn moves working except captures, en-pesant and promotions

+++ src/chess/move_square.rs
@@ -7,7 +7,7 @@ pub struct Move {
 
 #[derive(PartialEq, Copy, Clone, Debug)]
 pub struct Square {
-    pub rank: u8,
+    pub rank: i32, // included in (1..=8)
     pub file: char,
 }
 
@@ -39,7 +39,7 @@ impl Move {
 const BASE_TEN: u32 = 10;
 
 impl Square {
-    pub fn new(file: char, rank: u8) -> Option<Self> {
+    pub fn new(file: char, rank: i32) -> Option<Self> {
         if ('a'..='h').contains(&file) && (1..=8).contains(&rank) {
             Some(Square { file: file, rank: rank })
         } else {
@@ -50,7 +50,7 @@ impl Square {
     fn from_str(s: &str) -> Option<Self> {
         let mut chars= s.chars();
         let file = chars.next()?; // ? means if its None, the function returns early. otherwise return the value inside option
-        let rank = chars.next()?.to_digit(BASE_TEN).map(|d| d as u8)?;
+        let rank = chars.next()?.to_digit(BASE_TEN).map(|d| d as i32)?;
 
         if ('a'..='h').contains(&file) && (1..=8).contains(&rank) {
             Some(Square { file: file, rank: rank })
@@ -59,10 +59,10 @@ impl Square {
         }
     }
 
-    pub fn offset(&self, file_offset: i8, rank_offset: i8) -> Option<Square> {
+    pub fn offset(&self, file_offset: i8, rank_offset: i32) -> Option<Square> {
         let new_file = (self.file as i8 + file_offset) as u8 as char;
-        let new_rank = self.rank as i8 + rank_offset;
+        let new_rank = self.rank as i32 + rank_offset;
 
-        Square::new(new_file, new_rank as u8)
+        Square::new(new_file, new_rank as i32)
     }
 }
